requirements.txt
fastmcp==2.13.1
ytmusicapi==1.11.2
